Remove LightRope and LightString, as they don't offer enough benefits
Repeated profiling and JVM inspection has shown that the benefits of not
copying the string content is balanced by the overhead for the objects
themselves, and no caching strategies or optimizations were able to help
here. String is just too optimized already. If we used very long strings
this could make more of a difference, but for our use case, it is not
better or oftentimes slightly worse (5% more RAM, 30%-100% more CPU).