Switch to EUPL 1.2 as project license
Note that this is still indirectly compatible with EPL-2.0, as explained
in the EPL-2.0 FAQ:
https://www.eclipse.org/legal/epl-2.0/faq.php#h.tci84nlsqpgw
The added benefits are the official translations and compatibility with
more medium and strong copyleft free software licenses such as the GPL
family. It is very likely that the future EUPL will include EPL-2.0
explicitly and maybe other medium copyleft free software licenses.